feat(ui): remove control, use raw itemscontrol
