+++ src/Cobalt/App.axaml
@@ -4,7 +4,6 @@
              xmlns:fluentStyle="clr-namespace:FluentAvalonia.Styling;assembly=FluentAvalonia"
              xmlns:converters="clr-namespace:Cobalt.Converters"
              xmlns:system="clr-namespace:System;assembly=System.Runtime"
-             xmlns:controls="clr-namespace:Cobalt.Controls"
              RequestedThemeVariant="Default">
     <!-- "Default" ThemeVariant follows system theme variant. "Dark" or "Light" are other available options. -->
 
@@ -18,7 +17,7 @@
         <converters:IsEqualConverter x:Key="IsEqualConverter" />
         <converters:IsNotEqualConverter x:Key="IsNotEqualConverter" />
         <converters:ReminderDescriptionConverter x:Key="ReminderDescriptionConverter" />
-        <controls:PercentToLeftConverter x:Key="PercentToLeftConverter" />
+        <converters:PercentToLeftConverter x:Key="PercentToLeftConverter" />
 
         <!--!Font Awesome Free 6.5.1 by @fontawesome - https://fontawesome.com License - https://fontawesome.com/license/free Copyright 2023 Fonticons, Inc. Icon: bolt-->
         <Geometry x:Key="KillPath">M349.4 44.6c5.9-13.7 1.5-29.7-10.6-38.5s-28.6-8-39.9 1.8l-256 224c-10 8.8-13.6 22.9-8.9 35.3S50.7 288 64 288H175.5L98.6 467.4c-5.9 13.7-1.5 29.7 10.6 38.5s28.6 8 39.9-1.8l256-224c10-8.8 13.6-22.9 8.9-35.3s-16.6-20.7-30-20.7H272.5L349.4 44.6z</Geometry>
